Refactored the log dumper, use it in pilot tests to get more info
