+++ tests/e2e/framework/testInfo.go
@@ -28,18 +28,12 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/hashicorp/go-multierror"
 
 	"istio.io/istio/pkg/log"
 	"istio.io/istio/tests/util"
 )
 
 var (
-	resources = []string{
-		"pod",
-		"service",
-		"ingress",
-	}
 	testLogsPath = flag.String("test_logs_path", "", "Local path to store logs in")
 )
 
@@ -105,76 +99,7 @@ func (t testInfo) Update(r int) error {
 }
 
 func (t testInfo) FetchAndSaveClusterLogs(namespace string) error {
-	var multiErr error
-	fetchAndWrite := func(pod string) error {
-		cmd := fmt.Sprintf(
-			"kubectl get pods -n %s %s -o jsonpath={.spec.containers[*].name}", namespace, pod)
-		containersString, err := util.Shell(cmd)
-		if err != nil {
-			return err
-		}
-		containers := strings.Split(containersString, " ")
-		for _, container := range containers {
-			filePath := filepath.Join(t.TempDir, fmt.Sprintf("%s_container:%s.log", pod, container))
-			f, err := os.Create(filePath)
-			if err != nil {
-				return err
-			}
-			defer func() {
-				if err = f.Close(); err != nil {
-					log.Warnf("Error during closing file: %v\n", err)
-				}
-			}()
-			dump, err := util.ShellMuteOutput(
-				fmt.Sprintf("kubectl logs %s -n %s -c %s", pod, namespace, container))
-			if err != nil {
-				return err
-			}
-			if _, err = f.WriteString(fmt.Sprintf("%s\n", dump)); err != nil {
-				return err
-			}
-		}
-		return nil
-	}
-
-	_, err := util.Shell("kubectl get ingress --all-namespaces")
-	if err != nil {
-		return err
-	}
-	lines, err := util.Shell("kubectl get pods -n " + namespace)
-	if err != nil {
-		return err
-	}
-	pods := strings.Split(lines, "\n")
-	if len(pods) > 1 {
-		for _, line := range pods[1:] {
-			if idxEndOfPodName := strings.Index(line, " "); idxEndOfPodName > 0 {
-				pod := line[:idxEndOfPodName]
-				log.Infof("Fetching logs on %s", pod)
-				if err := fetchAndWrite(pod); err != nil {
-					multiErr = multierror.Append(multiErr, err)
-				}
-			}
-		}
-	}
-
-	for _, resrc := range resources {
-		log.Info(fmt.Sprintf("Fetching deployment info on %s\n", resrc))
-		filePath := filepath.Join(t.TempDir, fmt.Sprintf("%s.yaml", resrc))
-		if yaml, err0 := util.ShellMuteOutput(
-			fmt.Sprintf("kubectl get %s -n %s -o yaml", resrc, namespace)); err0 != nil {
-			multiErr = multierror.Append(multiErr, err0)
-		} else {
-			if f, err1 := os.Create(filePath); err1 != nil {
-				multiErr = multierror.Append(multiErr, err1)
-			} else {
-				if _, err2 := f.WriteString(fmt.Sprintf("%s\n", yaml)); err2 != nil {
-					multiErr = multierror.Append(multiErr, err2)
-				}
-			}
-		}
-	}
-	return multiErr
+	return util.FetchAndSaveClusterLogs(namespace, t.TempDir)
 }
 
 func (t testInfo) createStatusFile(r int) error {
